Move the Paths into path package.
- Add BukkitPaths class to create bukkit's path types.
